Polish utils.py, move some transformer to base_transformer
